pageserver: move gc_result into tenant
