+++ pageserver/src/lib.rs
@@ -20,7 +20,6 @@ pub use pageserver_api::keyspace;
 use tokio_util::sync::CancellationToken;
 mod assert_u64_eq_usize;
 pub mod aux_file;
-pub mod gc_result;
 pub mod metrics;
 pub mod page_cache;
 pub mod page_service;
